requirements.txt
reportlab
requests
discord-webhook
pytelegrambotapi
django
django-rest-framework
